Remove use of Listbox for  SearchTypeSelector on mobile
