--- src/components/Tabs/index.js
@@ -1,5 +1,0 @@
-import TabsPresentation from 'components/Tabs/TabsPresentation'
-
-export default {
-  Presentation: TabsPresentation,
-}
